BenchmarkCI.jl

BenchmarkCI — Module

BenchmarkCI

BenchmarkCI.judge: Run benchmarkpkg on target and baseline.

BenchmarkCI.postjudge: Post judgement as comment.

BenchmarkCI.displayjudgement: Print result of BenchmarkCI.judge.

BenchmarkCI.pushresult: Push benchmark result to a repository.

BenchmarkCI.judge — Function

judge()

Run benchmarkpkg on target and baseline.

Keyword Arguments

target :: Union{Nothing, AbstractString, BenchmarkConfig} = nothing: Benchmark target configuration. Default to the checked out working tree. A git commitish can be passed as a string. PkgBenchmark.BenchmarkConfig can be used for more detailed control (see PkgBenchmark.jl manual).

baseline :: Union{AbstractString, BenchmarkConfig} = "origin/master": Benchmark baseline configuration. See target.

pkgdir :: AbstractString = pwd(): Package root directory.

script :: AbstractString = "$pkgdir/benchmark/benchmarks.jl": The script that defines the SUITE global variable/constant.

project :: AbstractString = dirname(script): The project used to define and run benchmarks.

postprocess, retune, verbose, logger_factory: Passed to PkgBenchmark.benchmarkpkg.

BenchmarkCI.postjudge — Function

postjudge(; title = "Benchmark result")

Post judgement as a comment in the corresponding pull request.

BenchmarkCI.displayjudgement — Function

displayjudgement()

Print result of BenchmarkCI.judge.

BenchmarkCI.pushresult — Function

pushresult(; url, branch, sshkey, title)

Push benchmark result to branch in url.

Keyword Arguments

url::Union{AbstractString,Nothing} = nothing: Repository URL.

branch::AbstractString = "benchmark-results": Branch where the results are pushed.

sshkey::Union{AbstractString,Nothing} = nothing: Documenter.jl-style SSH private key (base64-encoded private key). Default to ENV["SSH_KEY"].

title::AbstractString = "Benchmark result": The title to be used in benchmark report.